Add files via upload
This is the final draft of my project 2 for BIMM 143 with Dr. Schiffer. In this project, I compare the amino acid sequences of four different spike proteins in order to determine if it is possible that the delta variant arose from another natural reservoir or through natural selection in the human population.